+++ packages/components/atoms/f-button/CHANGELOG.md
@@ -3,6 +3,20 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v1.8.0
+------------------------------
+*June 17, 2021*
+
+### Changed
+- Percy to run desktop / mobile in separate specs
+
+v1.7.0
+------------------------------
+*June 15, 2021*
+
+### Added
+- Support for Vue Router link and tests
+
 
 v1.6.1
 ------------------------------